Move ControlWithError and ValidityIndicator to form primitives
