+++ packages/ui/src/form/primitives/control-with-error/index.ts
@@ -1,0 +1,1 @@
+export { ControlWithError } from './control-with-error';
