+++ packages/ui/src/form/primitives/control-with-error/types.ts
@@ -1,4 +1,4 @@
-import type { ComponentProps } from '../utils/types';
+import type { ComponentProps } from '../../../utils/types';
 
 /**
  * HTML elements that support the Constraint Validation API.
@@ -63,20 +63,4 @@ export type ControlWithErrorProps = Omit<
 		label?: React.ReactNode;
 		required?: boolean;
 	} >;
-};
-
-export type ValidityIndicatorProps = {
-	/**
-	 * The `id` to apply to the indicator element, so it can be associated
-	 * with the control via `aria-describedby`.
-	 */
-	id?: string;
-	/**
-	 * The validation status to indicate.
-	 */
-	type: 'validating' | 'valid' | 'invalid';
-	/**
-	 * The message to display next to the status icon.
-	 */
-	message?: string;
 };
