+++ packages/ui/src/form/primitives/index.ts
@@ -1,5 +1,6 @@
 export * as Autocomplete from './autocomplete';
 export { Checkbox } from './checkbox';
+export { ControlWithError } from './control-with-error';
 export * as Combobox from './combobox';
 export * as Field from './field';
 export * as Fieldset from './fieldset';
@@ -8,3 +9,4 @@ export { InputLayout } from './input-layout';
 export * as Select from './select';
 export { SearchableChipSelect } from './searchable-chip-select';
 export { Textarea } from './textarea';
+export { ValidityIndicator } from './validity-indicator';
