+++ packages/ui/src/form/primitives/validity-indicator/validity-indicator.tsx
@@ -1,8 +1,8 @@
 import clsx from 'clsx';
 import { error, published } from '@wordpress/icons';
-import { Icon } from '../icon';
-import { Spinner } from '../spinner';
-import defenseStyles from '../utils/css/global-css-defense.module.css';
+import { Icon } from '../../../icon';
+import { Spinner } from '../../../spinner';
+import defenseStyles from '../../../utils/css/global-css-defense.module.css';
 import styles from './style.module.css';
 import type { ValidityIndicatorProps } from './types';
